+++ src/mp_api/matproj.py
@@ -26,6 +26,7 @@ from mp_api.magnetism.client import MagnetismRester
 from mp_api.search.client import SearchRester
 from mp_api.robocrys.client import RobocrysRester
 from mp_api.molecules.client import MoleculesRester
+from mp_api.synthesis.client import SynthesisRester
 
 _DEPRECATION_WARNING = (
     "MPRester is being modernized. Please use the new method suggested and "
@@ -157,6 +158,7 @@ class MPRester:
             api_key=api_key, endpoint=endpoint, include_user_agent=include_user_agent
 
         )
+        self.synth = SynthesisRester(api_key=api_key, endpoint=endpoint, include_user_agent=include_user_agent)
 
     def __enter__(self):
         """
