+++ src/mp_api/synthesis/client.py
@@ -1,0 +1,8 @@
+from mp_api.core.client import BaseRester
+from mp_api.synthesis.models import SynthesisDoc
+
+
+class SynthesisRester(BaseRester):
+
+    suffix = "synthesis"
+    document_model = SynthesisDoc
